Refactored location & weather logic
